+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-## 0.6.0
-
 * **BREAKING**: feat: Remove argument packages fron loadFonts
   - feat: loadFonts now support custom icons font like material_symbols_icons
+* **BREAKING**: feat: loadFontsFromPackage
+  - feat: loading fonts from current or external packages is now supported by loadFonts
 * **BREAKING**: feat: Add keyboardName to WindowConfigData
   - fix: It's now possible to use custom devices as WindowConfigData variant
 * feat: awaitImages now support FadeInImage.